+++ RELEASE-NOTES.md
@@ -21,6 +21,7 @@
 * #403 BoolAccum - java.lang.ArrayIndexOutOfBoundsException 
 * #407 "min", "max" aggregating functions for StrExp
 * #416 Zip ByteSource(s)
+* #417 Filesystem ByteSource(s)
 * #425 Series.replace(int,T)
 * #426 All "Json.load(..)" and "Json.save(..)" methods must be static
 * #427 Type-specific versions of Exp.$val(..)